+++ app/src/main/AndroidManifest.xml
@@ -10,16 +10,16 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.EasyLinks">
         <activity
-            android:name=".LoginSignupActivity"
+            android:name=".Activites.LoginSignupActivity"
             android:exported="true" />
-        <activity android:name=".SplashActivity">
+        <activity android:name=".Activites.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".Activites.MainActivity"></activity>
     </application>
 
 </manifest>
