+++ app/src/main/java/com/gdsc_jss/easylinks/Activites/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.gdsc_jss.easylinks
+package com.gdsc_jss.easylinks.Activites
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.gdsc_jss.easylinks.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
