+++ app/src/main/java/com/gdsc_jss/easylinks/Activites/SplashActivity.kt
@@ -1,10 +1,11 @@
-package com.gdsc_jss.easylinks
+package com.gdsc_jss.easylinks.Activites
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import com.gdsc_jss.easylinks.R
 
 class SplashActivity : AppCompatActivity() {
     private val SPLASH_TIME = 1500
@@ -17,6 +18,5 @@ class SplashActivity : AppCompatActivity() {
            startActivity(intent)
             finish()
         }, SPLASH_TIME.toLong())
-
     }
 }
