+++ app/src/main/java/com/gdsc_jss/easylinks/Fragments/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.gdsc_jss.easylinks
+package com.gdsc_jss.easylinks.Fragments
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,35 +8,37 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import com.gdsc_jss.easylinks.Activites.MainActivity
+import com.gdsc_jss.easylinks.R
 
 
-class SignupFragment : Fragment() {
+class LoginFragment : Fragment() {
 
-    private lateinit var loginBtn : TextView
-    private lateinit var signup : Button
+    private lateinit var signUpBtn : TextView
+    private lateinit var loginBtn : Button
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_signup, container, false)
+        val view = inflater.inflate(R.layout.fragment_login, container, false)
 
-        loginBtn = view.findViewById(R.id.login)
-        signup = view.findViewById(R.id.btnCreate)
+        signUpBtn = view.findViewById(R.id.btnSignup)
+        loginBtn = view.findViewById(R.id.btnLogin)
+
 
         return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        loginBtn.setOnClickListener {
+        signUpBtn.setOnClickListener{
             activity?.supportFragmentManager?.beginTransaction()
-                ?.replace(R.id.fragment_container, LoginFragment())
+                ?.replace(R.id.fragment_container, SignupFragment())
                 ?.commit()
         }
-        signup.setOnClickListener{
+        loginBtn.setOnClickListener{
             startActivity(Intent(requireContext(), MainActivity::class.java))
         }
     }
