+++ app/src/main/java/com/gdsc_jss/easylinks/Fragments/SearchFragment.kt
@@ -1,0 +1,19 @@
+package com.gdsc_jss.easylinks.Fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.gdsc_jss.easylinks.R
+
+class SearchFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_search, container, false)
+    }
+}
