+++ app/src/main/java/com/gdsc_jss/easylinks/Fragments/SignupFragment.kt
@@ -1,44 +1,44 @@
-package com.gdsc_jss.easylinks
+package com.gdsc_jss.easylinks.Fragments
 
 import android.content.Intent
 import android.os.Bundle
-import android.service.autofill.TextValueSanitizer
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
-import com.google.android.gms.dynamic.SupportFragmentWrapper
+import com.gdsc_jss.easylinks.Activites.MainActivity
+import com.gdsc_jss.easylinks.R
 
 
-class LoginFragment : Fragment() {
+class SignupFragment : Fragment() {
 
-    private lateinit var signUpBtn : TextView
-    private lateinit var loginBtn : Button
+    private lateinit var loginBtn : TextView
+    private lateinit var signup : Button
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_login, container, false)
+        val view = inflater.inflate(R.layout.fragment_signup, container, false)
 
-        signUpBtn = view.findViewById(R.id.btnSignup)
-        loginBtn = view.findViewById(R.id.btnLogin)
-
+        loginBtn = view.findViewById(R.id.login)
+        signup = view.findViewById(R.id.btnCreate)
 
         return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        signUpBtn.setOnClickListener{
+
+        loginBtn.setOnClickListener {
             activity?.supportFragmentManager?.beginTransaction()
-                ?.replace(R.id.fragment_container, SignupFragment())
+                ?.replace(R.id.fragment_container, LoginFragment())
                 ?.commit()
         }
-        loginBtn.setOnClickListener{
+        signup.setOnClickListener{
             startActivity(Intent(requireContext(), MainActivity::class.java))
         }
     }
