+++ app/src/main/res/layout/activity_login_signup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginSignupActivity">
+    tools:context=".Activites.LoginSignupActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
